refactor: reorganize authentication context and hooks for improved structure
